requirements.txt
langchain
langgraph
openai
tiktoken
google-api-python-client
python-dotenv
google-api-python-client
transformers
torch
youtube-transcript-api
langgraph
langchain-core
boto3
botocore
rich